Add support for copy-pasting cards between databases
The way it works is that selection of cards is copied to a in-memory database,
which acts as a "clipboard". It serves as staging area to view and edit changes
before putting them back onto a regular database. This database can also be
visualized in a special tab which can be opened from the File menu.

+++ src/gui/main_window.cpp
@@ -15,11 +15,12 @@
 #include <ui_main_window.h>
 
 #include "database_editor_widget.hpp"
+#include "sql_util.hpp"
 
 namespace
 {
 
-QString const SQL_DB_DRIVER("QSQLITE");
+QString const SQL_CLIPBOARD_CONN("CLIPBOARD");
 
 std::array const SQL_DATAS_TABLE_FIELDS = {
 	QString{"PRAGMA table_info('datas');"},
@@ -75,6 +76,13 @@ CREATE TABLE "texts" (
 )
 )");
 
+QString const SQL_QUERY_DATA_AND_TEXT_LIST(R"(
+SELECT datas.id,alias,setcode,type,atk,def,level,race,attribute,ot,category,
+texts.id,name,desc,str1,str2,str3,str4,str5,str6,str7,str8,str9,str10,str11,str12,str13,str14,str15,str16
+FROM datas, texts WHERE datas.id = texts.id AND texts.id IN (
+)");
+static constexpr int TEXT_INDEX_START = 11;
+
 } // namespace
 
 MainWindow::MainWindow(QWidget* parent)
@@ -93,6 +101,8 @@ MainWindow::MainWindow(QWidget* parent)
 	        &MainWindow::newDatabase);
 	connect(ui->actionOpenDatabase, &QAction::triggered, this,
 	        &MainWindow::openDatabase);
+	connect(ui->actionShowClipboardDatabase, &QAction::triggered, this,
+	        &MainWindow::showClipboardDatabase);
 	connect(ui->actionCloseDatabase, &QAction::triggered,
 	        [this]() { closeDatabase(-1); });
 	connect(ui->actionExit, &QAction::triggered, this, &MainWindow::close);
@@ -103,8 +113,16 @@ MainWindow::MainWindow(QWidget* parent)
 	        &MainWindow::saveData);
 	connect(ui->actionDeleteData, &QAction::triggered, this,
 	        &MainWindow::deleteData);
+	connect(ui->actionCopySelectedCards, &QAction::triggered, this,
+	        &MainWindow::copySelectedCards);
+	connect(ui->actionPasteClipboardCards, &QAction::triggered, this,
+	        &MainWindow::pasteClipboardCards);
 	connect(ui->actionHomepage, &QAction::triggered, this,
 	        &MainWindow::openHomepage);
+	// Setup "Clipboard" database
+	auto db = QSqlDatabase::addDatabase(SQL_DB_DRIVER, SQL_CLIPBOARD_CONN);
+	db.setDatabaseName(":memory:");
+	setupCleanDB(db);
 }
 
 MainWindow::~MainWindow()
@@ -141,6 +159,7 @@ void MainWindow::newDatabase()
 	QFile::remove(file);
 	auto db = QSqlDatabase::addDatabase(SQL_DB_DRIVER, file);
 	db.setDatabaseName(file);
+	setupCleanDB(db);
 	bool const isDbOpen = db.open();
 	Q_ASSERT(isDbOpen);
 	db.exec(SQL_QUERY_CREATE_DATAS_TABLE);
@@ -208,6 +227,23 @@ void MainWindow::openDatabase()
 	enableEditing(true);
 }
 
+void MainWindow::showClipboardDatabase()
+{
+	auto db = QSqlDatabase::database(SQL_CLIPBOARD_CONN, false);
+	auto const ptr = db.password();
+	if(!ptr.isEmpty())
+	{
+		ui->dbEditorTabsWidget->setCurrentWidget(&widgetFromConnection(ptr));
+		return;
+	}
+	// Create "Clipboard" view widget
+	auto* newTab =
+		new DatabaseEditorWidget(ui->dbEditorTabsWidget, SQL_CLIPBOARD_CONN);
+	ui->dbEditorTabsWidget->setCurrentIndex(
+		ui->dbEditorTabsWidget->addTab(newTab, tr("Clipboard")));
+	enableEditing(true);
+}
+
 void MainWindow::closeDatabase(int index)
 {
 	if(index < 0)
@@ -217,26 +253,62 @@ void MainWindow::closeDatabase(int index)
 	auto const dbConnection = tab->databaseConnection();
 	ui->dbEditorTabsWidget->removeTab(index);
 	delete tab;
-	QSqlDatabase::removeDatabase(dbConnection);
+	if(dbConnection != SQL_CLIPBOARD_CONN) // Long live the "Clipboard" db!
+		QSqlDatabase::removeDatabase(dbConnection);
+	else
+		QSqlDatabase::database(SQL_CLIPBOARD_CONN, false).setPassword("");
 	enableEditing(ui->dbEditorTabsWidget->count() != 0);
 }
 
 void MainWindow::newCard()
 {
-	static_cast<DatabaseEditorWidget*>(ui->dbEditorTabsWidget->currentWidget())
-		->newCard();
+	currentTab().newCard();
 }
 
 void MainWindow::saveData()
 {
-	static_cast<DatabaseEditorWidget*>(ui->dbEditorTabsWidget->currentWidget())
-		->saveData();
+	currentTab().saveData();
 }
 
 void MainWindow::deleteData()
 {
-	static_cast<DatabaseEditorWidget*>(ui->dbEditorTabsWidget->currentWidget())
-		->deleteData();
+	currentTab().deleteData();
+}
+
+void MainWindow::copySelectedCards()
+{
+	auto& tab = currentTab();
+	auto const codes = tab.selectedCards();
+	if(codes.size() == 0)
+		return;
+	auto dbSrc = QSqlDatabase::database(tab.databaseConnection(), false);
+	auto dbDst = QSqlDatabase::database(SQL_CLIPBOARD_CONN, false);
+	dbSrc.exec("DELETE FROM datas;");
+	dbSrc.exec("DELETE FROM texts;");
+	copyCards(codes, dbSrc, dbDst);
+	// TODO: Update target db's widget
+}
+
+void MainWindow::pasteClipboardCards()
+{
+	auto& tab = currentTab();
+	auto const tabDbConnection = tab.databaseConnection();
+	if(tabDbConnection == SQL_CLIPBOARD_CONN)
+		return;
+	auto dbSrc = QSqlDatabase::database(SQL_CLIPBOARD_CONN, false);
+	auto dbDst = QSqlDatabase::database(tabDbConnection, false);
+	// TODO: Confirm cards to be overwritten
+	auto const codes = [&]() -> QVector<quint32>
+	{
+		QVector<quint32> ret;
+		auto q = buildQuery(dbSrc, "SELECT id FROM datas;");
+		execQuery(q);
+		while(q.next())
+			ret.append(q.value(0).toUInt());
+		return ret;
+	}();
+	copyCards(codes, dbSrc, dbDst);
+	// TODO: Update target db's widget
 }
 
 void MainWindow::openHomepage()
@@ -247,16 +319,44 @@ void MainWindow::openHomepage()
 
 // private
 
-void MainWindow::enableEditing(bool editing)
+void MainWindow::copyCards(QVector<quint32> const& codes, QSqlDatabase& dbSrc,
+                           QSqlDatabase& dbDst)
 {
-	ui->actionSaveData->setEnabled(editing);
-	ui->actionNewCard->setEnabled(editing);
-	ui->actionDeleteData->setEnabled(editing);
-	ui->actionCloseDatabase->setEnabled(editing);
+	auto const stmt = [&]() -> QString
+	{
+		QString ret(SQL_QUERY_DATA_AND_TEXT_LIST);
+		for(auto const code : codes)
+			ret.append(QString::number(code)).append(',');
+		ret.append("0);");
+		return ret;
+	}();
+	auto q1 = buildQuery(dbSrc, stmt);
+	execQuery(q1);
+	auto const recordCount = q1.record().count();
+	auto q2 = buildQuery(
+		dbDst, QString(SQL_INSERT_DATA).replace("INSERT", "INSERT OR REPLACE"));
+	auto q3 = buildQuery(
+		dbDst, QString(SQL_INSERT_TEXT).replace("INSERT", "INSERT OR REPLACE"));
+	while(q1.next())
+	{
+		for(int i = 0; i < TEXT_INDEX_START; i++)
+			q2.bindValue(i, q1.value(i));
+		for(int i = 0; i < recordCount - TEXT_INDEX_START; i++)
+			q3.bindValue(i, q1.value(TEXT_INDEX_START + i));
+		execQuery(q2);
+		execQuery(q3);
+	}
+}
+
+DatabaseEditorWidget& MainWindow::currentTab() const
+{
+	Q_ASSERT(ui->dbEditorTabsWidget->count() > 0);
+	return *static_cast<DatabaseEditorWidget*>(
+		ui->dbEditorTabsWidget->currentWidget());
 }
 
 DatabaseEditorWidget& MainWindow::widgetFromConnection(
-	QString const& dbConnection)
+	QString const& dbConnection) const
 {
 	auto db = QSqlDatabase::database(dbConnection, false);
 	Q_ASSERT(!db.password().isEmpty());
@@ -264,4 +364,22 @@ DatabaseEditorWidget& MainWindow::widgetFromConnection(
 	auto const ptr = db.password().toULongLong(&ok, 16);
 	Q_ASSERT(ok);
 	return *reinterpret_cast<DatabaseEditorWidget*>(ptr);
+}
+
+void MainWindow::setupCleanDB(QSqlDatabase& db) const
+{
+	bool const isDbOpen = db.open();
+	Q_ASSERT(isDbOpen);
+	db.exec(SQL_QUERY_CREATE_DATAS_TABLE);
+	db.exec(SQL_QUERY_CREATE_TEXTS_TABLE);
+}
+
+void MainWindow::enableEditing(bool editing)
+{
+	ui->actionCloseDatabase->setEnabled(editing);
+	ui->actionSaveData->setEnabled(editing);
+	ui->actionNewCard->setEnabled(editing);
+	ui->actionDeleteData->setEnabled(editing);
+	ui->actionCopySelectedCards->setEnabled(editing);
+	ui->actionPasteClipboardCards->setEnabled(editing);
 }
